修复时间轴

+++ src/jmidi/Chord.java
@@ -5,20 +5,20 @@ package jmidi;
  */
 public class Chord {
 
-	public static int[][] get(int key) {
+	public static byte[][] get(int key) {
 		if (key == 6)
-			return new int[][] { { -2, 6 }, { -1, 3 }, { -1, 6 }, { -1, 7 }, { 0, 1 } };
+			return new byte[][] { { -2, 6 }, { -1, 3 }, { -1, 6 }, { -1, 7 }, { 0, 1 } };
 		if (key == 5)
-			return new int[][] { { -2, 5 }, { -1, 2 }, { -1, 5 }, { -1, 6 }, { -1, 7 } };
+			return new byte[][] { { -2, 5 }, { -1, 2 }, { -1, 5 }, { -1, 6 }, { -1, 7 } };
 		if (key == 4)
-			return new int[][] { { -2, 4 }, { -1, 1 }, { -1, 4 }, { -1, 5 }, { -1, 6 } };
+			return new byte[][] { { -2, 4 }, { -1, 1 }, { -1, 4 }, { -1, 5 }, { -1, 6 } };
 		if (key == 3)
-			return new int[][] { { -2, 3 }, { -2, 7 }, { -1, 3 }, { -1, 5 }, { -1, 7 } };
+			return new byte[][] { { -2, 3 }, { -2, 7 }, { -1, 3 }, { -1, 5 }, { -1, 7 } };
 		if (key == 2)
-			return new int[][] { { -2, 2 }, { -2, 6 }, { -1, 2 }, { -1, 4 }, { -1, 6 } };
+			return new byte[][] { { -2, 2 }, { -2, 6 }, { -1, 2 }, { -1, 4 }, { -1, 6 } };
 		if (key == 1)
-			return new int[][] { { -1, 1 }, { -1, 5 }, { 0, 1 }, { 0, 2 }, { 0, 3 } };
+			return new byte[][] { { -1, 1 }, { -1, 5 }, { 0, 1 }, { 0, 2 }, { 0, 3 } };
 
-		return new int[0][];
+		return new byte[0][];
 	}
 }
