+++ src/jmidi/Path.java
@@ -7,7 +7,7 @@ package jmidi;
  */
 public class Path {
 	
-	private static final int[][] data = new int[][]{
+	private static final byte[][] data = new byte[][]{
 		{ 4, 5, 3, 6 },
 		{ 6, 4, 1, 5 },
 		{ 6, 5, 4, 3, 2, 3, 4, 5 },
@@ -16,18 +16,18 @@ public class Path {
 		{ 6, 4, 5, 3, 4, 5, 3, 6 },
 	};
 	
-	public static int[] get(int seed) {
+	public static byte[] get(int seed) {
 		if(seed < data.length)
 			return data[seed];
 		
-		return new int[0];
+		return new byte[0];
 	}
 	
 	public static int size(){
 		return data.length;
 	}
 	
-	public static int[] rand(){
+	public static byte[] rand(){
 		return get(Note.rand(data.length));
 	}
 }
