+++ src/jmidi/Rhythm.java
@@ -7,7 +7,7 @@ package jmidi;
  */
 public class Rhythm {
 
-	private static final int[][] data = new int[][] {
+	private static final byte[][] data = new byte[][] {
 		{ 1, 0, 1, 0, 1, 0, 1, 0 },
 		{ 1, 0, 1, 1, 0, 1, 0, 1 },
 		{ 1, 1, 1, 1, 0, 0, 1, 0 },
@@ -16,18 +16,18 @@ public class Rhythm {
 		{ 1, 0, 1, 1, 1, 1 },
 	};
 
-	public static int[] get(int type) {
+	public static byte[] get(int type) {
 		if(type < data.length)
 			return data[type];
 		
-		return new int[0];
+		return new byte[0];
 	}
 	
 	public static int size(){
 		return data.length;
 	}
 	
-	public static int[] rand(){
+	public static byte[] rand(){
 		return get(Note.rand(data.length));
 	}
 }
