+++ src/test/Auto.java
@@ -16,8 +16,8 @@ public class Auto {
 		int section = 1; // 当前第几小节
 		int prev = 9; // 上一个音符，初始化设置为9
 		int range = 15; // 随机生成的音符范围
-		int[] rhythm = Rhythm.get(Note.rand(Rhythm.size())); // 随机选择节奏型
-		int[] path = Path.get(Note.rand(Path.size())); // 随机选择走向
+		byte[] rhythm = Rhythm.get(Note.rand(Rhythm.size())); // 随机选择节奏型
+		byte[] path = Path.get(Note.rand(Path.size())); // 随机选择走向
 
 		while (true) {
 			for (int i = 0, chd = 0; i < rhythm.length; i++) {
@@ -37,8 +37,8 @@ public class Auto {
 				// 和弦区
 				{
 					if (rhythm[i] == 1) {
-						int[][] chords = Chord.get(path[section - 1]);
-						int[] chord = chords[chd++ % chords.length];
+						byte[][] chords = Chord.get(path[section - 1]);
+						byte[] chord = chords[chd++ % chords.length];
 						
 						play.chord(Note.key(chord[0], chord[1])); // 播放和弦
 					}
